+++ index.d.ts
@@ -70,7 +70,9 @@ export type {Entry} from './source/entry.d.ts';
 export type {Entries} from './source/entries.d.ts';
 export type {SetReturnType} from './source/set-return-type.d.ts';
 export type {SetParameterType} from './source/set-parameter-type.d.ts';
-export type {Overloads, OverloadParameters, OverloadReturnType} from './source/function-overloads.d.ts';
+export type {Overloads} from './source/overloads.d.ts';
+export type {OverloadParameters} from './source/overload-parameters.d.ts';
+export type {OverloadReturnType} from './source/overload-return-type.d.ts';
 export type {Asyncify} from './source/asyncify.d.ts';
 export type {Simplify} from './source/simplify.d.ts';
 export type {SimplifyDeep} from './source/simplify-deep.d.ts';
